Crate inu_rs

A redux-like state manager based on inu.

Example

   use inu_rs::*;

   use async_std::{stream::interval, task::{block_on, spawn}};
   use futures::{FutureExt, Stream, StreamExt};
   use std::time::Duration;
   use std::pin::Pin;

   #[derive(Copy, Clone, Debug)]
   struct MyState {
       count: i32,
   }

   #[derive(Debug, Clone)]
   enum MyActions {
       TimerTicked,
   }

   #[derive(Debug, Clone)]
   enum MyEffects {
       ScheduleTick(u64),
   }

   impl State for MyState {
       type Action = MyActions;
       type Effect = MyEffects;

       fn apply_action(&mut self, action: &Self::Action) {
           match action {
               MyActions::TimerTicked => self.count = self.count + 1,
           }
       }
       fn apply_effect(&self, effect: &Self::Effect) -> Pin<Box<dyn Stream<Item = Option<Self::Action>>>> {
           match effect {
               MyEffects::ScheduleTick(tick_interval) => {
                   let interval = interval(Duration::from_millis(*tick_interval));
                   let stream = interval.take(2).map(|_| Some(MyActions::TimerTicked));
                   Box::pin(stream)
               }
           }
       }
   }

   fn main() {
       block_on(async {
           let initial_state = MyState { count: 0 };
           let mut inu = Inu::new(initial_state);
           let inu_handle = inu.get_handle();

           let stopped = inu
               .subscribe()
               .await
               .filter(|state| futures::future::ready(state.count >= 2))
               .into_future()
               .then(|_| async { inu_handle.stop().await.unwrap() });

           let mut inu_handle = inu.get_handle();
           // You can dispatch actions and effects asynchronously from other tasks using the
           // handle.
           spawn(async move {
                inu_handle.dispatch(None, Some(MyEffects::ScheduleTick(1)))
                    .await
                    .unwrap();
           });

           futures::join! {inu.run(), stopped };

           assert_eq!(inu.get_state().await.count, 2);
       });
   }

Structs

Handle
The handle you can use to stop an Inu

Inu
The Inu state manager.

OneshotReceiver
A future for a value that will be provided by another asynchronous task.

OneshotSender
A means of transmitting a single value to another task.

Receiver
The receiving end of a bounded mpsc channel.

SendError
The error type for Senders used as Sinks.

Sender
The transmission end of a bounded mpsc channel.

UnboundedReceiver
The receiving end of an unbounded mpsc channel.

Traits

State
Implement this trait for your application's state.